refactor: add parse header and line to parser.c
