+++ assembler/parser.c
@@ -1,5 +1,24 @@
 #include "./op.h"
 
+FileHeader parse_header(FILE *input) {
+    FileHeader header = {"", "", 0};
+    char line[MAX_LINE_LENGTH];
+    int name_found = 0;
+    int comment_found = 0;
+    while (fgets(line, sizeof(line), input) && (!name_found || !comment_found)) {
+        if (strncmp(line, ".name", 5) == 0) {
+            sscanf(line, ".name \"%[^\"]\"", header.name);
+            name_found = 1;
+        } else if (strncmp(line, ".comment", 8) == 0) {
+            sscanf(line, ".comment \"%[^\"]\"", header.comment);
+            comment_found = 1;
+        }
+    }
+    rewind(input);
+
+    return header;
+}
+
 parsed_line_t parse_line(const char *line) {
     parsed_line_t parsedLine = {TOKEN_UNKNOWN, "", "", {""}, 0};
     const char *inputPtr = line;
